Updated playRound function with dom methods, added variables, and event listeners
Updated playRound to change the output in the html of what users see after a round. Added declareWinner function to handle how a winner is decided and triggers a box to appear on screen showing the winner. Added displayScore function to display the score after each played round. Removed playGame function since the game is trigger by the human choosing a choice.